+++ fusion/src/Fusion/python/bindings.cpp
@@ -5,5 +5,5 @@ namespace py = pybind11;
 
 PYBIND11_MODULE(fusion, m) {
   m.doc() = "Fusion Tensor module exposing Tensor<double> (for composition)";
-  bind_tensor<double>(m, "Tensor");
+  bind_tensor<float>(m, "Tensor");
 }
